Refactor client overview (#836)
* Add removal of client label

* Use consistent naming

* Add snackbar for client label removal

* Add labels trackByName

* Address review suggestions

* Add new tests

* Simplify store implementation of label removal tracking

* Show remove client label error

* Remove show remove label error

* Add basic drawer implementation

* Disable page scrolling on drawer open and remove unneeded info from client details

* Change button style

* Change url on client details opening

* Test drawer with client details

* Refactor the client overview into a new component

* Add suggested changes

* Use css variables

* Refactor URL routing logic

* Move drawer to client component

* Reenable closing

* Cleanup
