+++ grr/server/grr_response_server/gui/ui/components/client/routing.ts
@@ -1,5 +1,8 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
+
+import {ClientDetails} from '../client_details/client_details';
+
 import {Client} from './client';
 
 
@@ -7,8 +10,13 @@ import {Client} from './client';
  * Client details page route.
  */
 export const CLIENT_ROUTES: Routes = [
-  {path: 'clients/:id', component: Client},
-  {path: 'clients/:id/details', component: Client},
+  {
+    path: 'clients/:id',
+    component: Client,
+    children: [
+      {path: Client.CLIENT_DETAILS_ROUTE, component: Client},
+    ],
+  },
 ];
 
 @NgModule({
