+++ grr/server/grr_response_server/gui/ui/components/client_details/client_details.ng.html
@@ -106,7 +106,7 @@
     <div class="section-title mat-h3">CPU & Memory</div>
     <table class="section">
       <tr *ngIf="client.osInfo.machine">
-        <td class="description">CPU Architecture</td>
+        <td class="description">CPU architecture</td>
         <td class="value">{{client.osInfo.machine}}</td>
       </tr>
       <tr *ngIf="client.memorySize">
