+++ grr/server/grr_response_server/gui/ui/components/client_details/client_details.ts
@@ -1,7 +1,5 @@
-import {ChangeDetectionStrategy, Component, OnDestroy, OnInit} from '@angular/core';
-import {ActivatedRoute} from '@angular/router';
-import {Subject} from 'rxjs';
-import {filter, map, takeUntil} from 'rxjs/operators';
+import {ChangeDetectionStrategy, Component} from '@angular/core';
+import {map} from 'rxjs/operators';
 
 import {ClientPageFacade} from '../../store/client_page_facade';
 import {getClientVersions} from './client_diff';
@@ -15,15 +13,11 @@ import {getClientVersions} from './client_diff';
   styleUrls: ['./client_details.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class ClientDetails implements OnInit, OnDestroy {
+export class ClientDetails {
   // Not static & private because is referenced in the template
   readonly INITIAL_NUM_USERS_SHOWN = 1;
   readonly INITIAL_NUM_INTERFACES_SHOWN = 3;
   readonly INITIAL_NUM_VOLUMES_SHOWN = 2;
-
-  private readonly id$ = this.route.paramMap.pipe(
-      map(params => params.get('id')),
-      filter((id): id is string => id !== null));
 
   readonly client$ = this.clientPageFacade.selectedClient$;
   readonly clientVersions$ = this.clientPageFacade.selectedClientVersions$.pipe(
@@ -35,18 +29,9 @@ export class ClientDetails implements OnInit, OnDestroy {
   currentNumInterfacesShown = this.INITIAL_NUM_INTERFACES_SHOWN;
   currentNumVolumesShown = this.INITIAL_NUM_VOLUMES_SHOWN;
 
-  private readonly unsubscribe$ = new Subject<void>();
-
   constructor(
-      private readonly route: ActivatedRoute,
       private readonly clientPageFacade: ClientPageFacade,
   ) {}
-
-  ngOnInit() {
-    this.id$.pipe(takeUntil(this.unsubscribe$)).subscribe(id => {
-      this.clientPageFacade.selectClient(id);
-    });
-  }
 
   getAccordionButtonState(
       totalNumElements: number, currentMaxNumElementsShown: number,
@@ -57,10 +42,5 @@ export class ClientDetails implements OnInit, OnDestroy {
       return 'no-button';
     }
     return 'show-less';
-  }
-
-  ngOnDestroy() {
-    this.unsubscribe$.next();
-    this.unsubscribe$.complete();
   }
 }
