+++ grr/server/grr_response_server/gui/ui/components/client_overview/client_overview.scss
@@ -1,0 +1,45 @@
+@use '~material-theme' as c;
+@use '~main' as main;
+
+$max-width: 1280px;
+$width: 80%;
+
+  .client-header {
+    box-sizing: border-box;
+    padding: 20px 0 30px 140px;
+    border-bottom: 1px solid c.mat-color(c.$foreground, divider-light);
+    height: main.$client-overview-height;
+
+    .fqdn-chips {
+      display: flex;
+      align-items: center;
+
+      h1 {
+        margin-bottom: 0;
+      }
+
+      .chips {
+        margin-left: 1em;
+        display: flex;
+        flex-wrap: wrap;
+
+        mat-chip {
+          min-height: 26px;
+          font-size: 12px;
+          border: 1px solid c.mat-color(c.$foreground, divider);
+          background-color: c.mat-color(c.$background, app-bar);
+        }
+
+        .add_label {
+          margin-left: 0.5em;
+          border-radius: 2em;
+        }
+      }
+    }
+
+    .os,
+    .client-id {
+      margin-bottom: 0.5em;
+    }
+  }
+
